First pass at Microlib architecture, CLI probably broken
