+++ .gitignore
@@ -1,3 +1,6 @@
+.DS_Store
+*~
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
