+++ metasub_utils/athena/constants.py
@@ -1,0 +1,6 @@
+"""Constants useful when working with the Athena Cluster at Weill Cornell Medical."""
+
+
+METASUB_RESULTS = '/athena/masonlab/scratch/projects/metagenomics/metasub/analysis/new_metasub_cap/.module_ultra/core_results'
+METASUB_DATA = '/athena/masonlab/scratch/projects/metagenomics/metasub/data'
+HALPHA_LIBRARY = '/athena/masonlab/scratch/projects/metagenomics/metasub/data/hudson_alpha_library'
