+++ metasub_utils/bridges/metasub_utils/bridges/constants.py
@@ -1,0 +1,9 @@
+"""Constants useful when working with the Bridges cluster at XSEDE-PSC."""
+
+from os.path import join, dirname
+
+
+SL_TABLE = join(dirname(__file__), 'datasuper_to_hauniq.tsv')
+
+METASUB_DATA = '/home/dcdanko/pylon5/MetaSUB'
+ASSEMBLIES = '/home/dcdanko/pylon5/metasub_assemblies'
