+++ metasub_utils/bridges/requirements.txt
@@ -1,0 +1,8 @@
+--index-url https://pypi.python.org/simple
+
+factory-boy==2.11.1
+
+../base/
+../microbe_census_data/
+
+-e .
