+++ metasub_utils/cli/metasub_utils/cli/cli.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+"""Console script for metasub_utils."""
+
+import click
+
+
+@click.group()
+def main(args=None):
+    """Console script for metasub_utils."""
+    pass
+
+
+if __name__ == "__main__":
+    main()
