+++ metasub_utils/metagenscope/metasub_utils/metagenscope/constants.py
@@ -1,0 +1,6 @@
+"""Constants for working with MetaGenScope."""
+
+
+MSUB_GROUP_NAME = 'MetaSUB'
+
+IDS = set(['ha_id', 'barcode', 'metasub_name', 'sl_name'])
