+++ metasub_utils/metagenscope/metasub_utils/metagenscope/utils.py
@@ -1,0 +1,9 @@
+"""Utilities for MetaGenScope."""
+
+
+def as_display_name(name):
+    """ Return a name with spaces and first letters capitalized."""
+    return ' '.join([
+        tkn[:1].upper() + tkn[1:]
+        for tkn in name.split('_')
+    ])
