--- metasub_utils/metasub_utils.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Main module."""
