+++ metasub_utils/wasabi/metasub_utils/wasabi/constants.py
@@ -1,0 +1,4 @@
+"""Constants for working with Wasabi Cloud Storage."""
+
+BUCKET_NAME = 'metasub'
+ENDPOINT_URL = 'https://s3.wasabisys.com'
