+++ metasub_utils/zurich/metasub_utils/zurich/constants.py
@@ -1,0 +1,5 @@
+"""Constants for working with the SFTP server at Zurich ETH."""
+
+URL = 'metasub.ethz.ch'
+DATA = '/data/'
+ASSEMBLIES = '/data/assemblies/'
